Refactor int to bytes
