+++ .integration/tests/dpc_testnet1.rs
@@ -140,7 +140,6 @@ fn test_testnet1_dpc_execute_constraints() {
         .add_output(Output::new(recipient.address, amount, Payload::default(), None).unwrap())
         .build(&mut rng)
         .unwrap();
-    let executables = state.executables();
 
     let authorization = DPC::<Testnet1Parameters>::authorize(&vec![], &state, &mut rng).unwrap();
 
@@ -149,7 +148,7 @@ fn test_testnet1_dpc_execute_constraints() {
 
     // Execute the programs.
     let mut executions = Vec::with_capacity(Testnet1Parameters::NUM_TOTAL_RECORDS);
-    for (i, executable) in executables.iter().enumerate() {
+    for (i, executable) in state.executables().iter().enumerate() {
         executions.push(executable.execute(i as u8, &local_data).unwrap());
     }
 
@@ -162,8 +161,8 @@ fn test_testnet1_dpc_execute_constraints() {
 
     let TransactionAuthorization {
         kernel,
-        input_records: old_records,
-        output_records: new_records,
+        input_records,
+        output_records,
         signatures,
         noop_compute_keys,
     } = authorization;
@@ -176,8 +175,8 @@ fn test_testnet1_dpc_execute_constraints() {
     // Generate the ledger membership witnesses
     let mut old_witnesses = Vec::with_capacity(Testnet1Parameters::NUM_INPUT_RECORDS);
 
-    // Compute the ledger membership witness and serial number from the old records.
-    for record in old_records.iter() {
+    // Compute the ledger membership witness and serial number from the input records.
+    for record in input_records.iter() {
         if record.is_dummy() {
             old_witnesses.push(MerklePath::default());
         } else {
@@ -198,11 +197,11 @@ fn test_testnet1_dpc_execute_constraints() {
     )
     .unwrap();
     let inner_private_variables = InnerPrivateVariables::new(
-        old_records.clone(),
+        input_records.clone(),
         old_witnesses,
         noop_compute_keys.iter().map(|key| key.clone().unwrap()).collect(), // This is safe only for this test case.
         signatures,
-        new_records.clone(),
+        output_records.clone(),
         encrypted_record_randomizers,
         program_randomness.clone(),
         local_data.leaf_randomizers().clone(),
@@ -228,7 +227,7 @@ fn test_testnet1_dpc_execute_constraints() {
     println!("=========================================================");
     let num_constraints = inner_circuit_cs.num_constraints();
     println!("Inner circuit num constraints: {:?}", num_constraints);
-    assert_eq!(332163, num_constraints);
+    assert_eq!(330571, num_constraints);
     println!("=========================================================");
 
     assert!(inner_circuit_cs.is_satisfied());
@@ -289,7 +288,7 @@ fn test_testnet1_dpc_execute_constraints() {
     println!("=========================================================");
     let num_constraints = outer_circuit_cs.num_constraints();
     println!("Outer circuit num constraints: {:?}", num_constraints);
-    assert_eq!(515990, num_constraints);
+    assert_eq!(523836, num_constraints);
     println!("=========================================================");
 
     assert!(outer_circuit_cs.is_satisfied());
